pipeline/workers/encoding/requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
appnope==0.1.3
asteroid-filterbanks==0.4.0
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
backcall==0.2.0
backports.cached-property==1.0.2
boto3==1.28.67
botocore==1.31.67
Brotli==1.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
colorlog==6.7.0
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
diart==0.7.0
docopt==0.6.2
einops==0.3.2
elevenlabs==0.2.26
exceptiongroup==1.1.3
executing==2.0.0
ffmpeg-python==0.2.0
filelock==3.12.4
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
future==0.18.3
google-auth==2.23.3
google-auth-oauthlib==1.0.0
greenlet==3.0.0
grpcio==1.59.0
hmmlearn==0.2.8
huggingface-hub==0.18.0
HyperPyYAML==1.2.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipython==8.16.1
jedi==0.19.1
jmespath==1.0.1
joblib==1.3.2
julius==0.2.7
kiwisolver==1.4.5
librosa==0.9.2
llvmlite==0.41.1
Mako==1.2.4
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mutagen==1.47.0
networkx==2.8.8
numba==0.58.1
numpy==1.26.1
oauthlib==3.2.2
omegaconf==2.3.0
openai==0.28.1
optuna==3.4.0
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
pooch==1.7.0
primePy==1.3
prompt-toolkit==3.0.39
protobuf==3.20.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.audio==2.1.1
pyannote.core==4.5
pyannote.database==4.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
pyDeprecate==0.3.2
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytorch-lightning==1.6.5
pytorch-metric-learning==1.7.3
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
rich==13.6.0
rsa==4.9
ruamel.yaml==0.17.39
ruamel.yaml.clib==0.2.8
Rx==3.2.0
s3transfer==0.7.0
scikit-learn==1.3.1
scipy==1.10.1
semver==2.13.0
sentencepiece==0.1.99
shellingham==1.5.3
simplejson==3.19.2
singledispatchmethod==1.0
six==1.16.0
sortedcontainers==2.4.0
sounddevice==0.4.6
SoundFile==0.10.3.post1
speechbrain==0.5.15
SQLAlchemy==2.0.22
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
temporalio==1.3.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
torch==1.13.1
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==0.13.1
torchmetrics==0.11.4
torchvision==0.14.1
tqdm==4.66.1
traitlets==5.11.2
typer==0.9.0
types-protobuf==4.24.0.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
wcwidth==0.2.8
websocket-client==1.6.4
websocket-server==0.6.4
websockets==11.0.3
Werkzeug==3.0.0
yarl==1.9.2
yt-dlp==2023.10.13
zipp==3.17.0
